@@ -1,9 +1,11 @@
-apply plugin: 'com.android.application'
-apply plugin: 'maven-publish'
+plugins {
+    id 'com.android.library'
+    id 'maven-publish'
+}
 
 repositories {
     mavenCentral()
-    //google()
+    google()
     maven { url "https://jitpack.io" }
 }
 
@@ -11,7 +13,7 @@ group = 'com.github.eatmyshortsbg'
 version = '1.0.1'
 
 android {
-    compileSdkVersion 30
+    compileSdkVersion 31
 
     defaultConfig {
         minSdkVersion 21
@@ -33,7 +35,8 @@ android {
 
 dependencies {
     compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:23.0.0'
+    compile 'androidx.appcompat:appcompat:1.4.1'
+    compile 'androidx.recyclerview:recyclerview:1.2.1'
 }
 
 afterEvaluate {
